[FEATURE] Tiny event logic for onStart, onSucces and onError
README updated with documentation for how to utilize this new feature. In short: you can configure a list of plugins which get executed when a plugin starts running, if an error occurs and when it finishes (if no error occurred). And you can do so in infinite levels (okay, honestly, limited by maximum YAML indentation level), adding event listeners to event listeners and so on.